+++ src/controllers/socialmedia.controller.ts
@@ -1,6 +1,7 @@
+import { UserDocument } from './../models/user.schema';
 import { Request, Response, NextFunction } from "express";
 import SocialMediaService from "../services/socialmedia.service";
-import { UserDocument } from "../models/user.schema";
+// import { UserDocument } from "../models/user.schema";
 // import { SocialMediaDocument } from "../models/socialMedia.schema";
 
 // Extend the existing Request interface to include the user property
@@ -11,6 +12,8 @@ interface AuthenticatedRequest extends Request {
 class SocialMediaController {
     static async createSocialMedia(req: AuthenticatedRequest, res: Response, next: NextFunction) {
         try {
+            const user = res.locals.UserDocument
+            req.body.userId = user._id
             const socialMedia = await SocialMediaService.createSocialMedia(req.body);
             res.json(socialMedia);
         } catch (error) {
@@ -20,8 +23,8 @@ class SocialMediaController {
 
     static async fetchUserSocialMedia(req: AuthenticatedRequest, res: Response, next: NextFunction) {
         try {
-            // Assuming the authenticated user object is stored in req.user
-            const user = req.user;
+            // Assuming the authenticated user object is stored in res.locals.user
+            const user = res.locals.user;
             if (!user) {
                 return res.status(401).json({ message: "Unauthorized" });
             }
@@ -37,6 +40,13 @@ class SocialMediaController {
 
     static async updateSocialMedia(req: AuthenticatedRequest, res: Response, next: NextFunction) {
         try {
+            const user = res.locals.user
+            const id = req.params.id
+            const existingSocialMedia = await SocialMediaService.findOne({ _id: id, userId: user._id })
+            if (!existingSocialMedia) {
+                return res.status(404).json({ message: "SocialMedia not found" })
+            }
+
             const updatedSocialMedia = await SocialMediaService.updateSocialMedia(req.params.id, req.body);
             if (!updatedSocialMedia) {
                 return res.status(404).json({ message: "SocialMedia not found" });
@@ -49,6 +59,12 @@ class SocialMediaController {
 
     static async deleteSocialMedia(req: AuthenticatedRequest, res: Response, next: NextFunction) {
         try {
+            const user = res.locals.user
+            const id = req.params.id
+            const existingProject = await SocialMediaService.findOne({ _id: id, userId: user._id })
+            if (!existingProject) {
+                return res.status(404).json({ message: "Project not found" })
+            }
             const deletedSocialMedia = await SocialMediaService.deleteSocialMedia(req.params.id);
             if (!deletedSocialMedia) {
                 return res.status(404).json({ message: "SocialMedia not found" });
